requirements.txt
openai
langchain
faiss-cpu
chromadb
tiktoken
unstructured
pandas